Add while loop for non-empty nodes array
While number of nodes in selection is greater than zero repeat already
known cycle of detecting if building is circle/orthogonal.

Script needs to be updated to go throw the array and decide the building
in the first iteration, go throw array and create buildings in the
second iteration and finally create residential area.