Fix perpendicular direction index
In 360 degrees there is 4 perpendicular directions. The length of
`DIRECTION` array containing all possible directions is `8`. Therefore,
if perpendicular direction is needed, `DIRECTION` array is looked up for
index increased by `8 / 4 = 2`. To avoid overflow `% DIRECTION.length`
is used.

Direction is increased by the half of the perpendicular direction
`8 / 4 / 2 = 8 / 8 = 1` if the center of the building is wanted.

This commit fix *perpendicular direction index* in the meaning of
scalability. In the future it is possible to change `DIRECTION` array
and increase the number of directions so hard-coded values are not
perspective.