Expand bulding border nodes from building center
Expanding was done from the center of the village. This patch introduces
expanding of residential area for each building from the middle of the
building by `0.00006` GPS coordinates points - approximately 6m. The
shape of the resulting residential area looks better. Also the algorithm
now does not need to call `expand` function.